Bug #33517637: Speed up microbenchmarks in debug mode
Running microbenchmarks in debug mode is pointless, but since we run
microbenchmarks as regular tests (because we don't have the full Google
microbenchmark framework, and haven't made a separate registration
framework), at least 100+100 iterations of each test (and minimum one
second) are run anyway. This wastes time when running unit tests (both
for people and on automated builds), so reduce the number of iterations
to a minimal number. This way, we can still check that the test doesn't
crash, but we don't need to wait for it.

Change-Id: I34e6bd6078ef0ef6384474e0d525bd461f116be8