Bug#33157531 'Improper Validation of Array Index'
Follow up patch, improve the hash value generation for 'i-pointers'

The i-pointers are known to contain 'patterns' of same page_no being used for
several objects, and as the objects are of same size, the offset component
is not uniformly distributed either.

Patch introduce a hashKey function based on the Murmur3 hash key
generation.

Reviewed By: Mauritz Sundell <mauritz.sundell@oracle.com>

Change-Id: Ifa668e702a4f8da5d734926183a7a3f0c22e1983