Bug#33516898: BIGINT converted to UNSIGNED_BIGINT with boolean operation
Bug#33117410 changed the resolved type of a boolean expression
from signed integer to unsigned bigint, in order to simplify the
type handling and facilitate another bugfix. However, what appeared
as a harmless metadata change turned out to be a more serious issue
for various connectors.

This patch will revert the metadata change. It will also fix the
actual problem solved by that change in a slightly different manner,
by adjusting "max_length" for a negation of an integer to at least
two characters.

No test case is required, the metadata decision is sufficiently
documented in the test cases.

Change-Id: Ifb5e98bdef347cbddef0be082e13cd8eecaebb48