Bug#33364732: Preliminary fixes for WL#14673 [bit fields]
Use bit fields for some of the members of AccessPath so that there is
room for adding a new member without increasing the size of the paths.
When we later add support for optimizing DELETE statements with the
hypergraph optimizer, a new member will be added for tracking which
table can be deleted from with so-called "immediate deletes".

Change-Id: Ie9df70f4a2b36b701a12480c4ed7e84f02e40061