Bug #33531039: Stop slow microbenchmarks after calibration
Before running a microbenchmark, we run it for 100 iterations to find
out how many iterations we need for a target time of approximately one
second. However, if a microbenchmark is slow, it will need more than
that second, and at that point, there's no point in running 100 more
iterations for the “real” run; we can just reuse the measurement from
the calibration.

Of course, the real Google microbenchmark framework has better routines
for this, but this workaround at least halves the time spent on running
such slow microbenchmarks.

Change-Id: Id155ead6fea99c9ebbdbe73c7c559cc1756be04d