Bug#33364732: Preliminary fixes for WL#14673 [bit fields]
Post-push fix for a build failure with some old versions of GCC:

access_path.h:291:27: error: 'AccessPath::safe_for_rowid' is too small
to hold all values of 'enum AccessPath::Safety' [-Werror]

This is caused by a bug in GCC causing false warnings if some cases
when an enum is used as a bit field. See
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=61414. The bug is fixed
in GCC 8.4 and GCC 9.3.

Work around the bug by not specifying AccessPath::safe_for_rowid as a
bit field. Instead, move it earlier so that it doesn't cause padding
between two other bit fields. It still wastes some bits (uses 8 bits
instead of the 2 it needs), but we can revisit that when we actually
need those bits.

Change-Id: I3766d7017c707fad089190639b9fd50b0496c9fc