Bug#33534747: perfschema.dml_handler fails intermittently with the hypergraph optimizer
The test creates and populates a temporary table using a CREATE TABLE
AS SELECT statement with no ORDER BY clause, so the order of the rows
in the table is non-deterministic. It occasionally inserts the rows in
a different order than what's recorded in the result file, which makes
the test print a diff and fail.

Fixed by adding an ORDER BY clause to make the order deterministic.

Change-Id: I8e4c46f5d333c5738218bc253224462e2916ccab