Bug#33305975: replica fails to read table map events with table name larger than 64 bytes
When receiving events from the source, the replica was checking if the
table name and database names were bigger than 64 bytes in size.

But the limitation in the server is that the number of characters is
not bigger than 64, meaning that valid names with characters that make
the size be above 64 bytes would case replication failures.

We simply remove the check as it is not up to the event to check logic
restrictions like this, just that the data is valid.

We do update the event to allow bigger database and table names for
the foreseeable future.

ReviewBoard: 27160