Bug#19947842 ASSERTION `HOST_LENGTH <= 60' FAILED IN SET_SETUP_ACTOR_KEY
When originally reported,
the bug was that inserting data into table performance_schema.setup_actors
could trigger an assert in debug, related to the data length.

In mysql when reported:
- the hostname in mysqld was extended to 255 ASCII characters (255 bytes)
- the `HOST` column in table setup_actors had room for only 60 bytes.

This has been fixed already since the HOST column was extended to 255
characters, but a similar issue still exists today:

When inserting data in the ROLE column, an assert can fail due to size:
- the ROLE DDL is declared as 32 UTF8MB3 characters,
  which can use up to 32*3 = 96 bytes
- the ROLE data storage is only 32 bytes in class PFS_role_name.

The fix is to declare ROLENAME_LENGTH as 32 * SYSTEM_CHARSET_MBMAXLEN,
instead of just 32.

Change-Id: I58b7b77c5fb4c15d8123661fa4fcd6f480474613