Bug#33142135: Assert for fields not marked as READ [bugfix,noclose]
IS NULL forces the evaluation of the "DEFAULT(v2)/1"
expression while fixing the arguments because the
division sets is_nullable = true(division by zero).
But the fields used by the DEFAULT() function are not marked
as READ.
The fix is to always mark the Item_default_value's field as READ
since the field values are already prepared as part of query
optimization so there is no reason to "mark" them.

Change-Id: I472039ddf7acdf1ee3a9da37659ff110ede87fc9