BUG#33438903: lv.plugin_running_mutex should be refactored to a Checkable_rwlock
The mutex `lv.plugin_running_mutex` was introduced to control
concurrent access between:
 1) START GROUP_REPLICATION;
 2) STOP GROUP_REPLICATION;
 3) SET @@GLOBAL.group_replication_XXX on options that cannot be
    run concurrently with START and STOP GROUP_REPLICATION;
 4) member actions UDFs;
on which the access was exclusive, thence the reason to use a mutex.
Later on, it was used to control:
 5) SELECT on performance_schema.replication_group_member_stats
though this only needs to be exclusive with START and STOP
GROUP_REPLICATION.

This is causing `SET @@GLOBAL.group_replication_XXX` commands to
fail when they do execute concurrently with `SELECT on
performance_schema.replication_group_member_stats`, which is
incorrect.

To solve the above issue this mutex was refactored into a read/write
lock, on which we have:
 1) START GROUP_REPLICATION;
      acquires a write lock.
 2) STOP GROUP_REPLICATION;
      acquires a write lock.
 3) SET @@GLOBAL.group_replication_XXX on options that cannot be
    run concurrently with START and STOP GROUP_REPLICATION;
      tries to acquire a read lock, if not possible the command ends
      immediately.
 4) member actions UDFs;
      tries to acquire a read lock, if not possible the command ends
      immediately.
 5) SELECT on performance_schema.replication_group_member_stats
      acquires a read lock.

Additional testing was added to cover all SET
@@GLOBAL.group_replication_XXX options and member actions UDFs
running concurrently with START and STOP GROUP_REPLICATION;

RB: 27277