Bug#33511690: MySQL fails by mysql_list_processes
The problem is due to bad check of current query block / query result
structure when calling mysqld_list_processes() using a direct call
from a client program.

It is rectified by adding an argument to mysqld_list_processes() that
tells whether it is called from a command object that manages
termination of sending to client, or whether the function terminates
sending by itself.

A comment on why it is required to have two different eof calls is
also added.

Change-Id: I71340d57d492c6c09d53a1e494d546b48731738b