WL#14672: Enable the hypergraph optimizer for UPDATE [3/8, access path]
Create an access path for UPDATE operations. The access path carries
information about which tables are updated and which tables can be
updated immediately while scanning the result of the join. This is
needed by the hypergraph optimizer in order to take the cost of
materializing row IDs for delayed updates into account.

The information stored in the access path is not yet used by the old
optimizer, except for printing in EXPLAIN FORMAT=TREE. There is also
no corresponding iterator for the new access path type yet. This will
be added in a later patch.

Change-Id: I81eb2146ad17153e6e6832e22e5416bf89f7a955