WL#14672: Enable the hypergraph optimizer for UPDATE [1/8, immediate update]
Factor out the code for determining which table (if any) can be
updated on the fly while scanning the result of the join. This is to
enable the creation of an access path for the UPDATE operation in a
later patch.

Note that this code is for the old optimizer only. It determines
whether or not a table can be updated on the fly after the join order
has been decided. The hypergraph optimizer will instead determine this
while constructing the plan, and prefer plans that can perform
immediate updates while scanning the result of the join.

Change-Id: Idfde29ce769695c00c19fab954eabb178eef088e