Bug #33541222 	my_abort is not producing core files after bug#31504206
After the `Bug #31504206 	ABORT() IS NOT THREAD-SAFE; CAUSES INTERRUPTION OF HANDLE_FATAL_SIGNAL`
the core file is not produced, if the `my_abort()` (and thus `my_server_abort()`) is called.
We added there a feature to not print stacktrace in the `handle_fatal_signal()` handler if it was just printed manually in `my_server_abort()`. Still, the `my_server_abort()` specifically requests `print_fatal_signal()` to not include core file, which causes the core file to not be created by any path.

We decouple writing core from `print_fatal_signal()` by moving the core writing block directly to `handle_fatal_signal()`

RB#27253