Bug #33538461 	InnoDB: Use ut:unique_ptr for global page_cleaner and its slots
This change was created while fixing some Valgrind issues after `Bug #33460092`.
In the end it showed up not necessary.
However, after the `ut::make_unique` was added recently, there is still no way to easily specify the type for a field that would be allocated using it, and this patch improve that.

RB#27243