Bug #33569770: Stackless CreateIteratorFromAccessPath()
Rewrite CreateIteratorFromAccessPath() so that it is non-recursive, and
thus needs a bounded amount of stack no matter the system.

The access path trees can be pretty deep, and the stack frames
can be big on certain compilers/setups, so instead of explicit
recursion, we now push jobs onto a MEM_ROOT-backed stack.
This uses up a little more RAM (the MEM_ROOT typically lives to the
end of the query), but reduces the stack usage greatly.

Change-Id: I0043808feecd1f9585a9ffc7ff383ad3c801752f