Bug#33574408: heap-use-after-free with multi-table UPDATE and view with WITH CHECK OPTION
Multi-table UPDATE statements may under some circumstances add
elements to a list each time it is executed. It never removes elements
from this list. This can make the memory footprint grow for each
execution of the statement, and make it read memory that was freed at
the end of an earlier execution.

Fixed by clearing the list at the end of execution.

Change-Id: If84fe2014e6d0c6f198f99e01c5da705d6efc066