Bug#33189767 - Improve logging of GCS communication issues
Problem
------------------------
I had an issue with a gr member not being able to join the group.

The logging shows:
2021-07-22T11:26:33.913088Z 0 [ERROR] [MY-011735] [Repl] Plugin
group_replication reported: '[GCS] Error connecting to the local group
communication engine instance.'

This error is far from clear: it does not specify what the error is, and it
does not specify what endpoint it was connecting to.

Improving the message would help.

I dived into the code and see one example of possible problems are TLS
issues, in my case this was due to a firewall issue. In either case reporting
the error in more detail would have helped diagnose and resolve the issue
faster.

The same error message is also reported in 2 different place.

Analysis and Fix
------------------------
The output of any of these two messages happens whenever we have an error:
- Connecting to the local XCom instance, which is done in order to be able
  to signal XCom that it has some work to do coming from GCS;
- Testing the connectivity of XCom, which is something that we do in order to
  fail-fast if any configuration is wrong.

The current messages are somehow cryptic and non-actionable, meaning that
looking at them on a log won't gives a clue of what to do.

We have improved the messages and made them specific for each supported
operating systems, thus easing the DBA role on the task of figuring out what
is wrong when it has an issue around this area.

RB: 27239