Bug#33579528 build fails with openssl 1.0.2 on windows
Issue
=====

build with openssl 1.0.2 on windows leads to:

  Building Custom Rule
  .../router/src/openssl/src/CMakeLists.txt
    openssl_plugin.cc
  .../shared\ws2def.h(240,25): error C2011:
  'sockaddr': 'struct' type redefinition

This is a problem in openssl 1.0.x which is specific to windows.

It can be avoided by including windows.h and winsock2.h before including
the openssl headers.

In this specific case, the openssl-include headers don't have to be
included as the code doesn't call openssl functions directly, but
through the TlsLibraryContext().

Change
======

- remove the openssl/*.h includes as the tls_context.h already does the
  right include order.

RB: 27351