Bug#34547467 ndb.ndb_multi_mgmd uses a fixed port
Problem:
An extra ndb_mgmd was started on fixed port 6000, which is out of the
allocated port range for the mtr worker thread.  This port could in
potentially be in conflict by another application running on same host,
or another mtr run on same host although this is unlikely since the
ndb_mgmd only is up for a short time (unless shutdown fails).

Fix:
Add a ndb_mgmd to test case config such that test uses two by mtr
started ndb_mgmd instead of starting one manually.

This now allows:

  $ ./mtr --parallel=4 ndb.ndb_multi_mgmd ndb.ndb_multi_mgmd

Change-Id: I187c73742c5f52c344b9ae37d504d0a4b8e5e8a4