BUG#33531967: server failure during group member recovery
When a member joins a Group Replication group, it goes through a
recovery step on which it fetches the data is missing in order to
have the same data as the group.
That step starts with randomly select a group member to be the donor
of the data transfer.
That random selection relies on `std::random_device`, which does
depend on the operating system and on the compiler. If this random
device is not available a exception is thrown by its constructor.
The Group Replication recovery implementation was not considering
that `std::random_device` could throw exceptions which was in
consequence resulting on the member joining server process to abort.

To solve the above issue, Group Replication recovery implementation
was updated to use `std::default_random_engine`, which is always
available.

RB: 27365