WL#14370, allow AT_MULTI_UNIQUE_KEY to be pushed
Remove restriction that a AT_MULTI_UNIQUE_KEY could not be pushed as
a pushed join root.

Note: For all practical purposes AT_MULTI_UNIQUE_KEY is similar to
AT_MULTI_PRIMARY_KEY, which has always been allowed to be a root.
Not allowing AT_MULTI_UNIQUE_KEY seems to be an oversight leftover
from early days of SPJ

Change-Id: I095ba88c5583b7940a0e944260c90ac0167b7659