WL#14370, Use '!m_scan_operations.contain(tab_no)' to detect lookup_operation
Replace test for ndbcluster_is_lookup_operation() with
m_scan_operations.contain(tab_no). Currently a mix of these two
methods for checking if we have a scan operation or not is used.

It is more efficient to do more of these checks with using the
m_scan_operations bitmap instead, and made code more readable as well.

Change-Id: I01a55e85dd6bd373c6c20354cf3db9b776da2403