BUG#32523769 TEST INNODB_S_GAP_DATA_LOCKS FAIL SPORADICALLY
Before this fix, the test perfschema.innodb_s_gap_data_locks
had spurious failures.

The root cause is that column LOCK_DATA in table
performance_schema.data_locks, in case of locks
taken in the innodb storage engine, can be:
- either a real value, when the underlying record is still present in the
  innodb buffer pool
- or NULL, when the underlying record is expelled from the innodb buffer
  pool.

This behavior is by design and not a bug in innodb,
so the table performance_schema.data_locks works as intended.

The test has been relaxed to pass when DATA_LOCK is NULL,
to avoid spurious failures.

Change-Id: I665f34bed0d578adc9ff345ab8427351fc410fe7