Bug #33592812: Clarify the description of WalkAccessPaths()
The doxygen comment for WalkAccessPaths() suggested that it would stop
the traversal when the `func` argument returned true. What it actually
does, in pre-order traversal, is to skip the rest of the subtree
rooted in the access path for which `func` returned true. In
post-order traversal, the return value of `func` is effectively
ignored, since the whole subtree has already been traversed when
`func` is called, so it's too late to skip it.

This patch updates the comment so that it describes the actual
behaviour. It also changes some wording that implies the
`cross_query_blocks` parameter is boolean (it used to be, but it is
now an enum). And it adds a unit test that verifies the current
behaviour, since code has started relying on it.

Change-Id: I444db661f97bce40b627363fb732c5914edd3953