WL#14370, Remove Table_access::dbug_print().
Can't remember when, if at all, it was ever usefull.

Would have needed a completely new implementation if it should
have been kept when we transition to use AccessPath instead of QEP_TAB
to analyse the optimized query plan. Is probably not worth the
effort, so we simply remove it.

Change-Id: I1e61914bfddfed4b744362b7d08f3e9c87cf9565