BUG#33184321: stop_all_network_providers: unsafe mix of type 'bool' and type 'int'
Description:
------------------
Building with Visual Studio, we get warning

network_provider_manager.cc(104)
warning C4805 '&=' unsafe mix of type 'bool' and type 'int'

There seems to something wrong with aggregation of status/return value.

Analysis and Fix:
-------------------
Both start() and stop() from the interface Network_provider were designed
returning an int value, in order to allow future-proof implementation. But in
practice, one is always returning either true or 0. This causes incoherence
in some parts of the implementation that uses this interface.

In order to keep it future-proof, but at the same time, coherent, one will
change start() and stop() return value to an std::pair<bool, int>, in which
bool will state the success of the operation and int can be used to return
an error code, if needed.

RB: 27188