Bug#31468581 - INVALID PID IN UNIX SOCKET LOCK FILE
Description
===========
X Plugin has its own "lock" file format, to be able to detect if it was
reserved by classic protocol or x protocol. Both classic and
X unix-socket handlers kills processes that wrote the "lockfile", thus
the mechanism was introduced to minimize chance to kill "wrong" process.

When user configured the classic protocol unix-socket file, to the same
value as it was assigned to X unix-socket, then in that case he was forced
to remove manually the X lock-file.

Fix
===

The file format of X unix socket lock files was aligned to classic mainly
to make easier fixing wrong configuration of `mysqld` when there is no
direct access to the host.

RB: 27395
Reviewed-by: Tomasz Stepniak <tomasz.s.stepniak@oracle.com>