Bug#33148961 FAILURE TO UPGRADE FROM 5.7, INVALID UTF8 CHARACTER STRING
PROBLEM:
--------
During server upgrade from 5.7 to 8.0, an invalid comment
string on a FIELD or within the body of a STORED ROUTINE
failed with an error message. This error provided
insufficient context to locate the object containing the
invalid comment string.

Additionally, An invalid comment string on a TABLE, INDEX,
TABLESPACE, PARTITION or SUBPARTITION caused an assertion
failure when attempting to store the string in the data
dictionary. This happened during CREATE/ALTER table as well.

ANALYSIS:
---------
Comment strings for these objects were not validated in 5.7,
and were stored as-is in the FRM file. This validation
function added in 8.0 throws an error during upgrade of the
FRM file with the invalid comment string.

It is not practical to guess the character set of an invalid
comment string during 8.0 upgrade and continue without
halting. Further, it is preferred to disallow invalid
metadata to be stored in the data dictionary. Hence we
choose to fail upgrade providing a more suitable error
message.

During CREATE/ALTER, when client charset differs from system
charset, any string literal was converted by the parser
which replaced invalid characters with '?'. However when the
client charset matched, this conversion did not take place
and the invalid characters were unchecked prior to attempted
storage in the DD. This caused the assertion errors in case
of invalid comments on TABLE, INDEX, TABLESPACE, PARTITION
and SUBPARTITION.

FIX:
----
If an invalid comment string is detected during CREATE,
ALTER, or UPGRADE from 5.7, we throw new error messages
which provide sufficient contextual information to address
the failure.

Change-Id: I9e0962bd04248f9353a398c6d07385148ef49dea