Bug #33550360: Inconsistent row estimations in the hypergraph optimizer [noclose]
Fix so that WHERE predicates that are above outer joins are not
attempted pushed to sargable; this caused inconsistent row counts,
but is actually potentially a correctness problem in the future.

The reason why it isn't a direct problem right now, is that only
simple equality predicates are sargable (we don't support x=y OR NULL),
and those generally simplify outer joins into inner joins.
The only cases where it _doesn't_ happen is when outer-to-inner
join conversion was possible, but we failed to do it (which can happen
due to a pass ordering issue).

This is a rare issue; in the entire test suite, only one case in
main.join is helped by this patch, and that case has other issues
causing row inconsistencies.

Change-Id: I5c2dada4be39029c872001eaafa0f369de66636d