Bug#33631207 decouple routing from plugin-env
Issue
=====

MySQLRouting terminates as soon as the plugin asks it to finish.

Currently, it passes the plugin-env to each class it spawns, to allow
them to check if it is time to finish too.

That creates a dependency between all the classes and the harness
plugin-env.

Instead of passing down the plugin env, the MySQLRouting class should
encapsulate the "is_running()" and hide the plugin-env dependency.

Change
======

- added MySQLRouting::is_running()
- removed all calls for mysql_harness::is_running(env) by
  MySQLRouting::is_running()
- replaced routing_stopped() by is_running()

RB: 27433