BUG#33609089: server failure during member autorejoin
When a member is expelled from Group Replication, by default it will
try to automatically rejoin the group.
It was noticed a mishandle of the concurrency of the statement
`SELECT * FROM performance_schema.replication_group_member_stats`
and the rejoin procedure, which could cause a server failure.

To prevent the above issue, these two concurrent operations are now
properly handled.

RB: 27405