Bug #33609599: Hypergraph: Arguments to NeighborhoodCache are swapped
NeighborhoodCache::InitSearch() and NeighborhoodCache::Store() take
"full_neighborhood" as second argument and "neighborhood" as third
argument, but their callers pass "neighborhood" and
"full_neighborhood" in the opposite order.

Fixed by swapping the order of the parameter lists so that they match
how the functions are called. Also added asserts to verify that the
arguments are passed in the correct order.

Change-Id: I8bde1e2e51073e8f3a20cc5316cc0b8ef4b75fe0