Bug #33550360: Inconsistent row estimations in the hypergraph optimizer [noclose]
Do not attempt to push conditions left as WHERE conditions down into
being sargable at all; it makes no sense (if they were possible to
push all the way down into the base tables and be sargable,
they would not be left as WHERE conditions, but made into join
conditions).

This fixes the last known MTR test row inconsistency issue that is not
related to left join reordering or EQ_REF clamping (but note that some
commits will be done out-of-order).

Change-Id: I4851648ec3355f77cc2c8d281ec22401ceba2674