Bug #33550360: Inconsistent row estimations in the hypergraph optimizer [noclose]
If there's a unique index over a single field, and that field is
being used in an equality predicate of the type field = <anything>,
cap that predicate's selectivity to 1/N, like we do for field = field
already. This fixes all EQ_REF inconsistencies for single-part indexes,
leaving only multi-part indexes (and thus, we can tighten the checks).

Change-Id: Ice45d1f38d26a9735439c8f2d2c34590221e8013