WL#14370, fix get_last_sj_inner(), and align with comming changes.
The existing, QEP_TAB based, AQP implementation takes the first/last sj_inner
values directly from the QEP_TAB itself. For semi-joins using the
'Duplicate weedout' strategy, it turns out that the 'last_sj_inner'
is not set in the QEP_TAB.

ndb_pushed_builder_ctx::is_pushable_with_root() partly compensated for that
by basing the logic which build the sj_nest on detecting a change in the
first_sj_inner instead. However, the 'unnest' logic of the sj_nest still
depended on 'last_sj_inner'. Thus calling validate_join_nest() might be
missed for a 'Duplicate wedout' semi-join.

Patch fix get_last_sj_inner() such that it return a correct 'last'
even when it is not set in the QEP_TAB's. Note that this WL will
replace this implementation of get_last_sj_inner() with an Accesspath
based implementation. (Which does return first/last correctly)

is_pushable_with_root() can be simplified as a result of this.

Change-Id: Idd086d8a887203f940c724501daf353c1ab5f90c