Bug#33594652 FIXES TIMEOUT FAILURE OF NDB_RESTORE_CONV_PK_NEGATIVE TEST
The fix for Bug#32604161 introduced a mandatory barrier after the
completion of data restore regardless of the options supplied so
that any operations with ndb_apply_status happens after the data
is restored completely.
In some cases, this causes one thread to wait in the barrier and
a failure in another thread doesn't lead to a complete failure since
the main thread is waiting for the thread waiting in the barrier to
complete indefinitely which never results in cancellation of the
barrier.

The probability of some cases is reduced by moving the barrier to be
executed conditionally when it is required by the options supplied.
Also, fixes a log message.

Change-Id: I6ab109a74d6fca521df9497bfe4a35b7dff38970